Modify packaging and Separate methods
