--- src/main/java/org/slam/service/bookhistory/BookHistoryUpdateService.java
@@ -1,7 +1,0 @@
-package org.slam.service.bookhistory;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BookHistoryUpdateService {
-}
