+++ src/main/java/org/slam/service/history/HistoryUpdateService.java
@@ -5,65 +5,31 @@ import org.slam.dto.book.Book;
 import org.slam.dto.book.BookHistory;
 import org.slam.dto.book.BookStatus;
 import org.slam.dto.book.UserAnswer;
-import org.slam.dto.common.Paginator;
 import org.slam.mapper.book.BookUpdateMapper;
-import org.slam.mapper.book.HistoryMapper;
+import org.slam.mapper.history.HistoryMapper;
+import org.slam.mapper.history.HistoryUpdateMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
 @AllArgsConstructor
-public class HistoryService {
+public class HistoryUpdateService {
 
     private final BookUpdateMapper bookUpdateMapper;
     private final HistoryMapper historyMapper;
-
-    public Map<String, Object> findMatchStatusHistory(BookStatus status, Paginator paginator) {
-        var resultMap = new HashMap<String, Object>();
-        paginator.setTotal(historyMapper.findTotalCount(status, paginator));
-
-        resultMap.put("paginator", paginator);
-        resultMap.put("bookList", historyMapper.findMatchStatusHistory(status, paginator));
-
-        return resultMap;
-    }
-
-    public List<BookHistory> findHistoryByBookId(Long bookId, String username) {
-        return historyMapper.findHistoryByBookId(bookId, username);
-    }
-
-    public Book findHistoryDetailsByBookId(Long bookId, Paginator paginator) {
-        var details = historyMapper.findHistoryDetailsByBookId(bookId, paginator);
-        if (!details.getCreatedBy().equals(paginator.getUsername())) {
-            details.setHistories(
-                    details.getHistories().stream()
-                            .filter( h -> h.getRequestedUser().equals(paginator.getUsername()) )
-                            .collect(Collectors.toList())
-            );
-        }
-        return details;
-    }
-
-    public List<BookHistory> findBookRequestHistoryById(Long id, String username) {
-        return historyMapper.findBookRequestHistoryById(id, username);
-    }
+    private final HistoryUpdateMapper historyUpdateMapper;
 
     public int reservationRequest(Long id, String modifier) {
         return historyMapper.insertHistory(Book.builder().id(id).status(BookStatus.ON_RESERVED).modifiedBy(modifier).build());
     }
 
     public int cancelReservationRequest(Long id, String username) {
-        return historyMapper.cancelReservation(id, username);
+        return historyUpdateMapper.cancelReservation(id, username);
     }
 
     public int returnRequest(Long id, String modifier) {
-        return historyMapper.updateBookHistoryToReturnRequest(Book.builder().id(id).status(BookStatus.RETURN_REQUEST).modifiedBy(modifier).build());
+        return historyUpdateMapper.updateBookHistoryToReturnRequest(Book.builder().id(id).status(BookStatus.RETURN_REQUEST).modifiedBy(modifier).build());
     }
 
     public void updateBookHistory(Book book, String username) {
@@ -81,12 +47,12 @@ public class HistoryService {
         if (book.getOwnerAnswer() == UserAnswer.ACCEPT) {
             book.setStatus(BookStatus.ON_LOAN);
             var history = makeHistoryMatchStatus(book, BookStatus.ON_LOAN);
-            historyMapper.updateBookHistoryToOnLoan(history);
+            historyUpdateMapper.updateBookHistoryToOnLoan(history);
         } else {
             book.setStatus(BookStatus.AVAILABLE);
             var history = makeHistoryMatchStatus(book, BookStatus.REJECTED);
-            historyMapper.updateBookHistoryStatus(history);
-            historyMapper.updateBookHistoryOnReservedToWaitForResponse(book.getId());
+            historyUpdateMapper.updateBookHistoryStatus(history);
+            historyUpdateMapper.updateBookHistoryOnReservedToWaitForResponse(book.getId());
         }
         updateBookStatus(book, username);
     }
@@ -95,12 +61,12 @@ public class HistoryService {
         if (book.getOwnerAnswer() == UserAnswer.ACCEPT) {
             book.setStatus(BookStatus.AVAILABLE);
             var history = makeHistoryMatchStatus(book, BookStatus.ON_LOAN);
-            historyMapper.updateBookHistoryToOnLoan(history);
+            historyUpdateMapper.updateBookHistoryToOnLoan(history);
         } else {
             book.setStatus(BookStatus.ON_LOAN);
             var history = makeHistoryMatchStatus(book, BookStatus.REJECTED);
-            historyMapper.updateBookHistoryStatus(history);
-            historyMapper.updateBookHistoryOnReservedToWaitForResponse(book.getId());
+            historyUpdateMapper.updateBookHistoryStatus(history);
+            historyUpdateMapper.updateBookHistoryOnReservedToWaitForResponse(book.getId());
         }
     }
