+++ src/main/java/org/slam/web/rest/MyItemRestController.java
@@ -1,4 +1,4 @@
-package org.slam.web;
+package org.slam.web.rest;
 
 import lombok.AllArgsConstructor;
 import org.slam.dto.book.Book;
@@ -6,7 +6,7 @@ import org.slam.dto.book.BookHistory;
 import org.slam.dto.book.BookStatus;
 import org.slam.dto.common.Paginator;
 import org.slam.service.book.BookSelectService;
-import org.slam.service.history.HistoryService;
+import org.slam.service.history.HistorySelectService;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,11 +15,11 @@ import java.util.Map;
 
 @RestController
 @AllArgsConstructor
-@RequestMapping("/my-page")
-public class MyPageRestController {
+@RequestMapping("/my-items")
+public class MyItemRestController {
 
     private final BookSelectService bookSelectService;
-    private final HistoryService historyService;
+    private final HistorySelectService historySelectService;
 
     @GetMapping("/{status}")
     public Map<String, Object> findMyItems(@PathVariable String status, Authentication auth, @ModelAttribute("paginator") Paginator paginator) {
@@ -28,11 +28,11 @@ public class MyPageRestController {
             case "my-books" :
                 return bookSelectService.findBookListByOwner(paginator);
             case "on-loan" :
-                return historyService.findMatchStatusHistory(BookStatus.ON_LOAN, paginator);
+                return historySelectService.findMatchStatusHistory(BookStatus.ON_LOAN, paginator);
             case "on-apply" :
-                return historyService.findMatchStatusHistory(BookStatus.WAIT_FOR_RESPONSE, paginator);
+                return historySelectService.findMatchStatusHistory(BookStatus.WAIT_FOR_RESPONSE, paginator);
             case "on-resv" :
-                return historyService.findMatchStatusHistory(BookStatus.ON_RESERVED, paginator);
+                return historySelectService.findMatchStatusHistory(BookStatus.ON_RESERVED, paginator);
             default:
                 return null;
         }
@@ -40,11 +40,11 @@ public class MyPageRestController {
 
     @GetMapping("/{id}/history")
     public List<BookHistory> findBookRequestHistoryById(@PathVariable Long id, Authentication auth) {
-        return historyService.findBookRequestHistoryById(id, auth.getName());
+        return historySelectService.findBookRequestHistoryById(id, auth.getName());
     }
 
     @PatchMapping("/{id}/history")
-    public String updateBookHistory(Book book, Authentication auth) {
+    public String updateItemHistory(Book book, Authentication auth) {
         System.out.println(book);
 //        historyService.updateBookHistory(book, auth.getName());
         return "success";
