+++ src/main/resources/mapper/book/HistoryMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.slam.mapper.book.HistoryMapper">
+<mapper namespace="org.slam.mapper.history.HistoryMapper">
 
     <select id="findTotalCount" resultType="Integer">
         SELECT /* historyMapper.findTotalCount */
